(#545) Rename AdvancedChocolateyDialog to AdvancedInstallView
