+++ Source/ChocolateyGui.Common.Windows/ChocolateyGui.Common.Windows.csproj
@@ -235,9 +235,6 @@
     <Compile Include="Commands\RelayCommand.cs" />
     <Compile Include="Commands\RoutedCommandBinding.cs" />
     <Compile Include="Commands\RoutedCommandMonitor.cs" />
-    <Compile Include="Controls\Dialogs\AdvancedChocolateyDialog.xaml.cs">
-      <DependentUpon>AdvancedChocolateyDialog.xaml</DependentUpon>
-    </Compile>
     <Compile Include="Controls\Dialogs\ChocolateyDialog.xaml.cs">
       <DependentUpon>ChocolateyDialog.xaml</DependentUpon>
     </Compile>
@@ -293,7 +290,7 @@
     <Compile Include="Utilities\PackageAuthorsComparer.cs" />
     <Compile Include="Utilities\ResourceReader.cs" />
     <Compile Include="ViewModels\AboutViewModel.cs" />
-    <Compile Include="ViewModels\Items\AdvancedInstallViewModel.cs" />
+    <Compile Include="ViewModels\AdvancedInstallViewModel.cs" />
     <Compile Include="ViewModels\Items\PackageViewModel.cs" />
     <Compile Include="ViewModels\Items\SourceViewModel.cs" />
     <Compile Include="ViewModels\LocalSourceViewModel.cs" />
@@ -306,6 +303,9 @@
     <Compile Include="ViewModels\SourcesViewModel.cs" />
     <Compile Include="Views\AboutView.xaml.cs">
       <DependentUpon>AboutView.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Views\AdvancedInstallView.xaml.cs">
+      <DependentUpon>AdvancedInstallView.xaml</DependentUpon>
     </Compile>
     <Compile Include="Views\LocalSourceView.xaml.cs">
       <DependentUpon>LocalSourceView.xaml</DependentUpon>
@@ -338,10 +338,6 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <Page Include="Controls\Dialogs\AdvancedChocolateyDialog.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
     <Page Include="Controls\Dialogs\ChocolateyDialog.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -379,6 +375,10 @@
       <SubType>Designer</SubType>
     </Page>
     <Page Include="Views\AboutView.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Page Include="Views\AdvancedInstallView.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
