+++ Source/ChocolateyGui.Common.Windows/Views/AdvancedInstallView.xaml
@@ -1,14 +1,14 @@
-﻿<UserControl x:Class="ChocolateyGui.Common.Windows.Controls.Dialogs.AdvancedChocolateyDialog" 
+﻿<UserControl x:Class="ChocolateyGui.Common.Windows.Views.AdvancedInstallView" 
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" 
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:items="clr-namespace:ChocolateyGui.Common.Windows.ViewModels.Items" 
              xmlns:mah="http://metro.mahapps.com/winfx/xaml/controls" 
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:properties="clr-namespace:ChocolateyGui.Common.Properties;assembly=ChocolateyGui.Common" 
+             xmlns:properties="clr-namespace:ChocolateyGui.Common.Properties;assembly=ChocolateyGui.Common"
+             xmlns:viewModels="clr-namespace:ChocolateyGui.Common.Windows.ViewModels"
              MaxHeight="500" 
              MaxWidth="700" 
-             d:DataContext="{d:DesignInstance items:AdvancedInstallViewModel}" 
+             d:DataContext="{d:DesignInstance viewModels:AdvancedInstallViewModel}" 
              d:DesignHeight="450" 
              d:DesignWidth="800" 
              mc:Ignorable="d">
